refactor: move types from useIntelligentSearch to ila26/types
